pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "microbit-mcp"
version = "0.1.0"
description = "MCP Server for the micro:bit"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.12.0",
    "pyserial-asyncio>=0.6",
]

[project.scripts]
microbit-mcp = "mcp_server.server:cli_main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server"]

[tool.ruff]
exclude = ["src/microbit/main.py"]

